Add lombok to project. Refactor code. Encrypt application.properties password.
