+++ build.gradle
@@ -18,6 +18,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa:2.6.2'
     implementation 'io.swagger.core.v3:swagger-core:2.1.12'
     implementation 'io.springfox:springfox-boot-starter:3.0.0'
+    implementation 'com.github.ulisesbocchio:jasypt-spring-boot-starter:3.0.4'
+
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.13.0'
 
     testImplementation 'org.springframework.boot:spring-boot-starter-test:2.6.2'
