+++ src/main/java/ru/neoflex/accountservice/AccountServiceApplication.java
@@ -1,13 +1,16 @@
 package ru.neoflex.accountservice;
 
+import com.ulisesbocchio.jasyptspringboot.annotation.EnableEncryptableProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Configuration;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @SpringBootApplication
 @EnableScheduling
 @EnableSwagger2
+@EnableEncryptableProperties
 public class AccountServiceApplication {
 
     public static void main(String[] args) {
