+++ src/main/java/ru/neoflex/accountservice/entity/BankAccount.java
@@ -1,9 +1,12 @@
 package ru.neoflex.accountservice.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.*;
+import org.hibernate.Hibernate;
 import ru.neoflex.accountservice.model.enums.Sex;
 
 import javax.persistence.*;
+import java.util.Objects;
 import java.util.UUID;
 
 import static javax.persistence.EnumType.STRING;
@@ -11,6 +14,11 @@ import static javax.persistence.EnumType.STRING;
 @Entity
 @Table(name = "account")
 @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
+@Getter
+@Setter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
 public class BankAccount {
     @Id
     @Column(name = "account_id")
@@ -32,74 +40,16 @@ public class BankAccount {
     @Enumerated(STRING)
     private Sex sex;
 
-    public BankAccount(UUID uuid, String firstName, String middleName, String lastName, long accountNumber, Sex sex) {
-        this.uuid = uuid;
-        this.firstName = firstName;
-        this.middleName = middleName;
-        this.lastName = lastName;
-        this.accountNumber = accountNumber;
-        this.sex = sex;
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || Hibernate.getClass(this) != Hibernate.getClass(o)) return false;
+        BankAccount that = (BankAccount) o;
+        return uuid != null && Objects.equals(uuid, that.uuid);
     }
-
-    public BankAccount() {}
 
     @Override
-    public String toString() {
-        return "BankAccount{" +
-                "uuid=" + uuid +
-                ", firstName='" + firstName + '\'' +
-                ", middleName='" + middleName + '\'' +
-                ", lastName='" + lastName + '\'' +
-                ", accountNumber=" + accountNumber +
-                ", sex=" + sex +
-                '}';
-    }
-
-    public UUID getUuid() {
-        return uuid;
-    }
-
-    public void setUuid(UUID uuid) {
-        this.uuid = uuid;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getMiddleName() {
-        return middleName;
-    }
-
-    public void setMiddleName(String middleName) {
-        this.middleName = middleName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public long getAccountNumber() {
-        return accountNumber;
-    }
-
-    public void setAccountNumber(long accountNumber) {
-        this.accountNumber = accountNumber;
-    }
-
-    public Sex getSex() {
-        return sex;
-    }
-
-    public void setSex(Sex sex) {
-        this.sex = sex;
+    public int hashCode() {
+        return getClass().hashCode();
     }
 }
