+++ src/main/java/ru/neoflex/accountservice/model/AddressDTO.java
@@ -2,6 +2,7 @@ package ru.neoflex.accountservice.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.*;
 
 import javax.persistence.Enumerated;
 import java.util.Objects;
@@ -9,6 +10,11 @@ import java.util.Objects;
 import static javax.persistence.EnumType.STRING;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
+@Setter
+@Getter
+@ToString
+@NoArgsConstructor
+@AllArgsConstructor
 public class AddressDTO {
 
     @JsonProperty(value = "city")
@@ -33,75 +39,6 @@ public class AddressDTO {
     @Enumerated(STRING)
     private String country;
 
-    public AddressDTO(String city, String streetName, String streetAddress, String buildingNumber, String zipCode, String postCode, String country) {
-        this.city = city;
-        this.streetName = streetName;
-        this.streetAddress = streetAddress;
-        this.buildingNumber = buildingNumber;
-        this.zipCode = zipCode;
-        this.postCode = postCode;
-        this.country = country;
-    }
-
-    public AddressDTO() {
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public String getStreetName() {
-        return streetName;
-    }
-
-    public String getStreetAddress() {
-        return streetAddress;
-    }
-
-    public String getBuildingNumber() {
-        return buildingNumber;
-    }
-
-    public String getZipCode() {
-        return zipCode;
-    }
-
-    public String getPostCode() {
-        return postCode;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public void setStreetName(String streetName) {
-        this.streetName = streetName;
-    }
-
-    public void setStreetAddress(String streetAddress) {
-        this.streetAddress = streetAddress;
-    }
-
-    public void setBuildingNumber(String buildingNumber) {
-        this.buildingNumber = buildingNumber;
-    }
-
-    public void setZipCode(String zipCode) {
-        this.zipCode = zipCode;
-    }
-
-    public void setPostCode(String postCode) {
-        this.postCode = postCode;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
-    }
-
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
@@ -113,18 +50,5 @@ public class AddressDTO {
     @Override
     public int hashCode() {
         return Objects.hash(getCity(), getStreetName(), getStreetAddress(), getBuildingNumber(), getZipCode(), getPostCode(), getCountry());
-    }
-
-    @Override
-    public String toString() {
-        return "AddressDTO{" +
-                "city='" + city + '\'' +
-                ", streetName='" + streetName + '\'' +
-                ", streetAddress='" + streetAddress + '\'' +
-                ", buildingNumber='" + buildingNumber + '\'' +
-                ", zipCode='" + zipCode + '\'' +
-                ", postCode='" + postCode + '\'' +
-                ", country='" + country + '\'' +
-                '}';
     }
 }
