+++ src/main/resources/application.properties
@@ -1,8 +1,9 @@
 server.port = 8082
-spring.datasource.password=123
 spring.datasource.url=jdbc:postgresql://localhost:5432/neoflex-project
 spring.datasource.username=postgres
-
+spring.datasource.password=ENC(bwjR7Lzh56m5Q+anV+c8m+kbI5DkKXwahx5zd0AxPGJyke5XoyHAFUt/PGDiUca/)
+jasypt.encryptor.iv-generator-classname=org.jasypt.iv.RandomIvGenerator
+jasypt.encryptor.algorithm=PBEWITHHMACSHA512ANDAES_256
+jasypt.encryptor.password=${IDEA_ADDRESS_SERVICE_ENCRYPTION}
 spring.jackson.serialization.fail-on-empty-beans=false
-
 spring.jpa.show-sql=false
